2816: Move generated protobuf code to `src-gen`
