+++ administration/buf.gen.yaml
@@ -4,7 +4,7 @@ version: v2
 
 plugins:
   - local: ../node_modules/.bin/protoc-gen-es
-    out: src/generated
+    out: src-gen
     opt:
       - target=ts
