+++ administration/src/cards/extensions/BavariaCardTypeExtension.tsx
@@ -2,7 +2,7 @@ import { FormControl, InputLabel, MenuItem, Select } from '@mui/material'
 import React, { ReactElement } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { BavariaCardType as BavariaCardTypeEnum } from '../../generated/card_pb'
+import { BavariaCardType as BavariaCardTypeEnum } from '../../card_pb'
 import type { Extension, ExtensionComponentProps } from './extensions'
 
 export const BAVARIA_CARD_TYPE_EXTENSION_NAME = 'bavariaCardType'
