+++ administration/src/routes/erstellen/components/CardSelfServiceActivation.test.tsx
@@ -2,10 +2,10 @@ import { create } from '@bufbuild/protobuf'
 import { fireEvent } from '@testing-library/react'
 import React from 'react'
 
+import { DynamicActivationCodeSchema } from '../../../card_pb'
 import { generateCardInfo, initializeCard } from '../../../cards/card'
 import { CreateCardsResult } from '../../../cards/createCards'
 import getCustomDeepLinkFromQrCode from '../../../cards/getCustomDeepLinkFromQrCode'
-import { DynamicActivationCodeSchema } from '../../../generated/card_pb'
 import { config } from '../../../project-configs/koblenz/config'
 import { CustomRenderOptions, renderWithOptions } from '../../../testing/render'
 import CardSelfServiceActivation from './CardSelfServiceActivation'
